UIRS-15 refactor form: filter submitted values outside (#44)
