+++ package.json
@@ -103,6 +103,7 @@
   },
   "peerDependencies": {
     "@folio/stripes": "^6.0.0",
+    "final-form": "^4.18.2",
     "react": "*",
     "react-intl": "^5.7.0 ",
     "react-router": "^5.2.0",
